+++ controllers/publisher.controller.js
@@ -39,6 +39,8 @@ export default class AdminController {
 
   async updatePublisher(req, res){
     try{
+      const publisher = await this.publisherService.getPublisher(req.params.id);
+      if (publisher===null) return res.sendStatus(404);
       const updatedPublisher = await this.publisherService.updatePublisher(req.body, req.params.id);
       res.send(updatedPublisher);
     } catch (error){
@@ -49,6 +51,8 @@ export default class AdminController {
 
   async deletePublisher(req, res){
     try{
+      const publisher = await this.publisherService.getPublisher(req.params.id);
+      if (publisher===null) return res.sendStatus(404);
       const deletedPublisher = await this.publisherService.deletePublisher(req.params.id);
       res.send("The publisher ("+deletedPublisher["name"]+") has been Deleted succesfully");
     } catch (error){
